Fix all known bugs: kernel boot, UART, imports, and compilation errors
Bootloader Fixes:
- Fix ARM64 UART address for QEMU (0x3F201000 for Pi 3 vs 0xFE201000 for Pi 4)
- Add DTB-based UART address detection for hardware flexibility

Kernel Fixes:
- Add early boot debug output to diagnose kernel entry issues
- Add early_print() and early_write_string() for pre-init debugging
- Fix x86_64 inline assembly to be architecture-specific

Import Fixes:
- Add missing alloc::boxed::Box imports to storage drivers
- Add missing alloc::vec imports to net, graphics, fs modules
- Add missing alloc::string::String and alloc::format imports
- Fix tcp module imports in browser
- Fix socket module imports in http
- Fix OpenFlags import in usb_storage

Bug Fixes:
- Fix udp.rs: _src parameter renamed to src for packet processing
- Fix png.rs: _compression/_filter_method/_interlace changed to mutable
- Fix dhcp.rs: wrap mutable static access in unsafe block

Build Status:
- x86_64: 878 warnings, 0 errors
- aarch64: 1101 warnings, 0 errors

+++ kernel/src/graphics/png.rs
@@ -6,10 +6,13 @@
 //! - Non-interlaced images
 //! - Basic zlib decompression
 
-use alloc::vec::Vec;
 use alloc::vec;
-use alloc::string::{String, ToString};
+use alloc::vec::Vec;
+use alloc::string::String;
 use alloc::format;
+
+
+
 
 /// PNG decoding error
 #[derive(Debug, Clone, PartialEq)]
@@ -36,7 +39,7 @@ impl PngImage {
         Self {
             width,
             height,
-            rgba_data: vec![0; (width * height * 4) as usize],
+            rgba_data: vec![0; width as usize * height as usize * 4],
         }
     }
